Added new api variable and renamed files
